move bar chart and stacked bar chart into a Charts folder that would contain all the charts
