+++ adobe-ourwork-aem-dev/components/Charts/BarChart/BarChart.type.ts
